requirements.txt
streamlit
openai
python-dotenv
Pillow
ollama
mistralai